requirements.txt
requests
feedparser
python-dotenv
beautifulsoup4
google-genai
instagrapi
playwright
